Per mile should be a whole number in pence

+++ db/migrate/20180205131119_create_pricing_rules.rb
@@ -3,7 +3,7 @@ class CreatePricingRules < ActiveRecord::Migration
     create_table :pricing_rules do |t|
       t.string :name
       t.integer :rule_type
-      t.float :per_mile, default: 0.0
+      t.integer :per_mile, default: 0
       t.jsonb :stages, null: false, default: {}
       t.float :child_multiplier, default: 0.5
       t.float :return_multiplier, default: 1.5
